Create forceApplyBuilding method for popup and dataGrid
