+++ js/ui/grid_core/ui.grid_core.columns_view.js
@@ -578,6 +578,11 @@ exports.ColumnsView = modules.View.inherit(columnStateMixin).inherit({
             result = [],
             $cells;
 
+        var forceApplyBindings = this.option("forceApplyBindings");
+        if(forceApplyBindings) {
+            forceApplyBindings();
+        }
+
         $tableElement = $tableElement || that._getTableElement();
 
         if($tableElement) {
